Remove check for plotly
A conditional always makes the logic more complex.  We don't really care if
we generate the additional markers and then ignore them.  That would only be
a problem if it consumed a lot of resources to do so.  Here is is very
cheap.  It's more important that the logic be simple.